test: move basics to ExportTests
