+++ Google.Cloud.Spanner.NHibernate.Tests/SpannerSchemaTests.cs
@@ -61,25 +61,6 @@ namespace Google.Cloud.Spanner.NHibernate.Tests
             // Configuration.AddMapping(mapping);
         }
         
-        [Fact]
-        public void Basics()
-        {
-            var conn = new SpannerRetriableConnection(new SpannerConnection(new SpannerConnectionStringBuilder(_fixture.ConnectionString, ChannelCredentials.Insecure)
-            {
-                EmulatorDetection = EmulatorDetection.None,
-            }));
-            _fixture.SpannerMock.AddOrUpdateStatementResult("Update singers", StatementResult.CreateUpdateCount(1L));
-            var cmd = conn.CreateDmlCommand("Update singers");
-            // var cmd = conn.CreateDdlCommand("CREATE TABLE Foo");
-            cmd.ExecuteNonQuery();
-
-            // var requests = _fixture.DatabaseAdminMock.Requests.OfType<UpdateDatabaseDdlRequest>();
-            // Assert.Collection(requests, request => Assert.Collection(request.Statements, statement => Assert.Equal("CREATE TABLE Foo", statement)));
-
-            var requests = _fixture.SpannerMock.Requests.OfType<ExecuteSqlRequest>();
-            Assert.Collection(requests, request => Assert.Equal("Update singers", request.Sql));
-        }
-        
 /*
         [Fact]
         public void SpannerExporterCanGenerateCreateModel()
